Add project level analysis job details
