+++ src/app/components/audio-analysis/audio-analysis.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from "@angular/core";
 import { RouterModule } from "@angular/router";
 import { getRouteConfigForPage } from "@helpers/page/pageRouting";
 import { SharedModule } from "@shared/shared.module";
-import { AudioAnalysisComponent } from "./pages/details/details.component";
+import { AudioAnalysisJobComponent } from "./pages/details/details.component";
 import { AudioAnalysesComponent } from "./pages/list/list.component";
 import { NewAudioAnalysisComponent } from "./pages/new/new.component";
 import { AudioAnalysisResultsComponent } from "./pages/results/results.component";
@@ -11,7 +11,7 @@ import { audioAnalysisRoutes } from "./audio-analysis.routes";
 const components = [
   AudioAnalysesComponent,
   NewAudioAnalysisComponent,
-  AudioAnalysisComponent,
+  AudioAnalysisJobComponent,
   AudioAnalysisResultsComponent,
 ];
