+++ src/app/components/audio-analysis/pages/details/details.component.html
@@ -1,0 +1,4 @@
+<div *ngIf="!failure">
+  <h1>Analysis Job Details</h1>
+  <baw-detail-view [fields]="fields" [model]="analysisJob"></baw-detail-view>
+</div>
